Move CodeSlider into setup folder
