+++ webview/src/setup/components/experiments/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../shared/variables';
 
 .codeSlider {
   margin-top: 40px;
